Improve module organization
